requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-telegram-bot==21.1.1
websockets==12.0
aiohttp==3.9.5